pyproject.toml
[project]
name = "medium-clone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "alembic>=1.16.5",
    "celery[redis]>=5.5.3",
    "email-validator>=2.3.0",
    "fastapi-mail>=1.5.0",
    "fastapi[all]>=0.116.1",
    "jinja2>=3.1.6",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.43",
    "starlette-admin>=0.15.1",
    "uvicorn[standard]>=0.35.0",
]
